feat: alert창 모달로 수정
