+++ src/components/Card/myActibityCardInfo/index.tsx
@@ -40,15 +40,15 @@ export default function MyActibitiyCardInfo({ data, delActivity }: { data: Activ
       {isDropShow && (
         <div className='absolute z-10 bottom-[-12.7rem] bg-white right-0 flex flex-col rounded-md border-[#DDD] border max-lg:bottom-[-11.7rem] max-md:bottom-[-9.3rem]'>
           <Link
-            href={`/myactivities/editactivity/${id}`}
-            className='flex justify-center items-center px-[4.6rem] py-[1.8rem] text-[1.8rem] text-[#4B4B4B] leading-[2.2rem] font-medium border-[#DDD] border-b max-lg:text-[1.6rem] max-lg:px-[3.4rem] max-lg:py-[1.4rem] max-md:text-[1.2rem] max-md:px-[2.8rem] max-md:py-[1rem]'
+            href={`/my/activities/editactivity/${id}`}
+            className='flex justify-center items-center px-[4.6rem] py-[1.8rem] text-[1.8rem] text-[#4B4B4B] hover:bg-gray-400 hover:text-white leading-[2.2rem] font-medium border-[#DDD] border-b max-lg:text-[1.6rem] max-lg:px-[3.4rem] max-lg:py-[1.4rem] max-md:text-[1.2rem] max-md:px-[2.8rem] max-md:py-[1rem]'
           >
             수정하기
           </Link>
           <button
             type='button'
             onClick={() => delActivity(id)}
-            className='flex justify-center items-center text-[1.8rem] text-[#4B4B4B] font-medium px-[4.6rem] leading-[2.2rem] py-[1.8rem] max-lg:text-[1.6rem] max-lg:px-[3.4rem] max-lg:py-[1.4rem] max-md:text-[1.2rem] max-md:px-[2.8rem] max-md:py-[1rem]'
+            className='flex justify-center items-center text-[1.8rem] text-[#4B4B4B] hover:bg-gray-400 hover:text-white font-medium px-[4.6rem] leading-[2.2rem] py-[1.8rem] max-lg:text-[1.6rem] max-lg:px-[3.4rem] max-lg:py-[1.4rem] max-md:text-[1.2rem] max-md:px-[2.8rem] max-md:py-[1rem]'
           >
             삭제하기
           </button>
