+++ src/pages/my/activities/index.tsx
@@ -61,7 +61,7 @@ export default function MyActivities() {
       <main className='bg-[#fafafa] mb-[12rem] min-h-[81rem] max-md:min-h-[67rem]'>
         <div className='flex justify-between mb-[2.4rem]'>
           <h2 className='text-[3.2rem] text-[#000] leading-[3.8rem] font-bold'>내 체험 관리</h2>
-          <Link href='/myactivities/postactivity' className='flex justify-center items-center w-[12rem] h-[4.8rem] text-[#fff] bg-[#112211] font-bold text-[1.6rem] leading-[2.6rem] rounded-[0.4rem]'>
+          <Link href='/my/activities/postactivity' className='flex justify-center items-center w-[12rem] h-[4.8rem] text-[#fff] bg-[#112211] font-bold text-[1.6rem] leading-[2.6rem] rounded-[0.4rem]'>
             체험 등록하기
           </Link>
         </div>
